+++ src/main/java/com/divroll/domino/repository/UserRepository.java
@@ -33,7 +33,8 @@ import java.util.List;
 public interface UserRepository {
     String createUser(String instance, String storeName, String username, String password, String[] read, String[] write);
 
-    boolean updateUser(String instance, String storeName, String entityId, String newUsername, String newPassword, String[] read, String[] write);
+    boolean updateUser(String instance, String storeName, String entityId,
+                       String newUsername, String newPassword, String[] read, String[] write);
 
     User getUser(String instance, String storeName, String userID);
 
@@ -41,5 +42,5 @@ public interface UserRepository {
 
     boolean deleteUser(String instance, String storeName, String userID);
 
-    List<User> listUsers(String instance, String storeName, long skip, long limit);
+    List<User> listUsers(String instance, String storeName, int skip, int limit);
 }
